refactor: ProfileInList
